requirements.txt
absl-py==2.1.0
accelerate==0.28.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
beautifulsoup4==4.12.3
blinker==1.7.0
blis==0.7.11
Brotli==1.1.0
browser-cookie3==0.19.1
bs4==0.0.2
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
cmake==3.29.0.1
codetiming==1.2.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.4
configparser==6.0.1
constants==0.6.0
contourpy==1.2.0
crayons==0.4.0
crfsuite==0.3.1
cryptography==42.0.5
cssselect==1.2.0
cycler==0.12.1
cymem==2.0.8
datasets==2.18.0
dateparser==1.2.0
debugpy==1.8.1
decorator==5.1.1
demjson3==3.0.6
dill==0.3.8
distro==1.9.0
EditorConfig==0.12.4
elastic-transport==8.13.0
elasticsearch==8.10.1
environs==9.5.0
et-xmlfile==1.1.0
evaluate==0.4.1
exceptiongroup==1.2.0
executing==2.0.1
facebook-page-scraper==5.0.2
facebook-scraper==0.2.59
fake-useragent==1.5.1
fastai==2.7.14
fastapi==0.104.0
fastcore==1.5.29
fastdownload==0.0.7
fastprogress==1.0.3
filelock==3.13.3
Flask==3.0.3
flatbuffers==24.3.7
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.2.0
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.1
h11==0.14.0
h2==4.1.0
h5py==3.10.0
hpack==4.0.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.0
humanfriendly==10.0
hyperframe==6.0.1
idna==3.6
importlib_metadata==7.1.0
importlib_resources==6.4.0
ipykernel==6.29.3
ipython==8.18.1
ipywidgets==8.1.1
itsdangerous==2.1.2
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.3
joblib==1.3.2
jsbeautifier==1.14.9
jsonlines==4.0.0
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab-widgets==3.0.9
kaitaistruct==0.10
keras==3.1.1
kiwisolver==1.4.5
langcodes==3.3.0
libclang==18.1.1
lit==18.1.2
loguru==0.5.3
lxml==5.1.0
lz4==4.3.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
milksnake==0.1.6
ml-dtypes==0.3.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
namex==0.0.7
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py==12.535.133
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
nvitop==1.3.2
onnx==1.16.0
onnxruntime==1.17.1
openai==1.14.3
openpyxl==3.1.2
opt-einsum==3.3.0
optimum==1.18.0
optree==0.11.0
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.1
parse==1.20.1
parso==0.8.3
pexpect==4.9.0
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pillow==10.2.0
platformdirs==4.2.0
praw==7.1.4
prawcore==1.5.0
preshed==3.0.9
pretty-errors==1.2.25
prettytable==3.10.0
progressbar2==4.4.2
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pushshift.py==0.1.2
pyarrow==15.0.2
pyarrow-hotfix==0.6
pyasn1==0.6.0
pycparser==2.21
pycryptodomex==3.20.0
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
pydash==7.0.6
pyee==11.1.0
Pygments==2.17.2
pyOpenSSL==24.1.0
pyparsing==3.1.2
pyppeteer==2.0.0
pyquery==2.0.0
PySocks==1.7.1
python-crfsuite==0.9.10
python-dateutil==2.8.2
python-dotenv==1.0.1
python-utils==3.8.2
pytorch-crf==0.7.2
pytz==2024.1
pyvi==0.1.1
PyYAML==6.0.1
pyzmq==25.1.2
regex==2023.12.25
requests==2.31.0
requests-html==0.10.0
responses==0.18.0
rich==13.7.1
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
selenium==4.1.0
selenium-wire==5.1.0
sentencepiece==0.2.0
seqeval==1.2.2
six==1.16.0
sklearn-crfsuite==0.3.6
smart-open==6.4.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
starlette==0.27.0
StrEnum==0.4.15
sympy==1.12
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
thinc==8.2.3
threadpoolctl==3.4.0
tokenizers==0.13.3
torch==2.0.1+cu117
torchaudio==2.0.2+cu117
TorchCRF==1.1.0
torchvision==0.15.2+cu117
tornado==6.4
tqdm==4.66.1
traitlets==5.14.2
-e git+https://github.com/datquocnguyen/transformers.git@8bb9b7e9219a2bdbecfcf18d1b8d8a63a5c189b5#egg=transformers
transformers-crf==0.3.0
trio==0.25.0
trio-websocket==0.11.1
triton==2.0.0
typer==0.3.2
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
update-checker==0.18.0
urllib3==2.0.7
urllib3-secure-extra==0.1.0
uvicorn==0.23.2
vaxiom-client @ http://minio.dev.ftech.ai/lib-packages/vaxiom_client-0.1.0rc8-py3-none-any.whl#sha256=0057672f201ba170b3403bb1741746c15b7c99b32bf5da84cba14063fa29c549
w3lib==2.1.2
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
webdriver-manager==3.2.2
webencodings==0.5.1
websocket-client==1.7.0
websockets==10.4
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.18.1
zstandard==0.22.0
